# Conflicts:
#	WindowsGSM/DiscordBot/Commands.cs
#	WindowsGSM/MainWindow.xaml.cs

+++ WindowsGSM/GameServer/MCBE.cs
@@ -7,6 +7,7 @@ using System.Text.RegularExpressions;
 using System.Windows;
 using System.Collections.Generic;
 using System;
+using WindowsGSM.Functions;
 
 namespace WindowsGSM.GameServer
 {
@@ -42,6 +43,8 @@ namespace WindowsGSM.GameServer
         public string Maxplayers = "10";
         public string Additional = string.Empty;
 
+        public string RegexString = @"https:\/\/www.minecraft\.net\/bedrockdedicatedserver\/bin-win\/(bedrock-server-(.*?)\.zip)";
+
         public MCBE(Functions.ServerConfig serverData)
         {
             _serverData = serverData;
@@ -58,7 +61,7 @@ namespace WindowsGSM.GameServer
                 configText = configText.Replace("{{max-players}}", Maxplayers);
                 string tempPort = _serverData.ServerPort;
                 configText = configText.Replace("{{server-port}}", tempPort);
-                configText = configText.Replace("{{server-portv6}}", (int.Parse(tempPort)+1).ToString());
+                configText = configText.Replace("{{server-portv6}}", (int.Parse(tempPort) + 1).ToString());
                 configText = configText.Replace("{{level-name}}", Defaultmap);
                 File.WriteAllText(configPath, configText);
             }
@@ -82,61 +85,62 @@ namespace WindowsGSM.GameServer
                 return null;
             }
 
-            Process p;
-            if (!AllowsEmbedConsole)
+            var p = new Process
             {
-                p = new Process
+                StartInfo =
                 {
-                    StartInfo =
-                    {
-                        WorkingDirectory = workingDir,
-                        FileName = exePath,
-                        WindowStyle = ProcessWindowStyle.Minimized,
-                    },
-                    EnableRaisingEvents = true
-                };
-                p.Start();
-            }
-            else
+                    CreateNoWindow = false,
+                    WorkingDirectory = ServerPath.GetServersServerFiles(_serverData.ServerID),
+                    FileName = exePath,
+                    WindowStyle = ProcessWindowStyle.Minimized,
+                    UseShellExecute = false,
+                },
+                EnableRaisingEvents = true
+            };
+
+            // Set up Redirect Input and Output to WindowsGSM Console if EmbedConsole is on
+            if (_serverData.EmbedConsole)
             {
-                p = new Process
-                {
-                    StartInfo =
-                    {
-                        WorkingDirectory = workingDir,
-                        FileName = exePath,
-                        WindowStyle = ProcessWindowStyle.Minimized,
-                        CreateNoWindow = true,
-                        UseShellExecute = false,
-                        RedirectStandardInput = true,
-                        RedirectStandardOutput = true,
-                        RedirectStandardError = true
-                    },
-                    EnableRaisingEvents = true
-                };
-                var serverConsole = new Functions.ServerConsole(_serverData.ServerID);
+                p.StartInfo.RedirectStandardInput = true;
+                p.StartInfo.RedirectStandardOutput = true;
+                p.StartInfo.RedirectStandardError = true;
+                p.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+                p.StartInfo.CreateNoWindow = true;
+                var serverConsole = new ServerConsole(_serverData.ServerID);
                 p.OutputDataReceived += serverConsole.AddOutput;
                 p.ErrorDataReceived += serverConsole.AddOutput;
-                p.Start();
-                p.BeginOutputReadLine();
-                p.BeginErrorReadLine();
             }
 
-            return p;
+            // Start Process
+            try
+            {
+                p.Start();
+                if (_serverData.EmbedConsole)
+                {
+                    p.BeginOutputReadLine();
+                    p.BeginErrorReadLine();
+                }
+                return p;
+            }
+            catch (Exception e)
+            {
+                Error = e.Message;
+                return null; // return null if fail to start
+            }
         }
 
         public async Task Stop(Process p)
         {
             await Task.Run(() =>
             {
-                if (p.StartInfo.RedirectStandardInput)
-                {
-                    p.StandardInput.WriteLine("stop");
-                }
-                else
-                {
-                    Functions.ServerConsole.SendMessageToMainWindow(p.MainWindowHandle, "stop");
-                }
+                Functions.ServerConsole.SetMainWindow(p.MainWindowHandle);
+                Functions.ServerConsole.SendWaitToMainWindow("stop");
+                Functions.ServerConsole.SendWaitToMainWindow("{ENTER}");
+                p.WaitForExit(4000);
+                p.CloseMainWindow();
+                p.WaitForExit(500);
+                if (!p.HasExited)
+                    p.Kill();
             });
         }
 
@@ -155,11 +159,12 @@ namespace WindowsGSM.GameServer
                 using (WebClient webClient = new MCBEWebclient())
                 {
                     string html = await webClient.DownloadStringTaskAsync("https://www.minecraft.net/en-us/download/server/bedrock/");
-                    Regex regex = new Regex(@"https:\/\/minecraft\.azureedge\.net\/bin-win\/(bedrock-server-(.*?)\.zip)");
+                    Regex regex = new Regex(RegexString);
                     var matches = regex.Matches(html);
 
                     if (matches.Count <= 0)
                     {
+                        Error = "could not find BedrockServer versions";
                         return null;
                     }
 
@@ -194,7 +199,7 @@ namespace WindowsGSM.GameServer
                     string remoteBuild = await GetRemoteBuild();
 
                     string html = await webClient.DownloadStringTaskAsync("https://www.minecraft.net/en-us/download/server/bedrock/");
-                    Regex regex = new Regex(@"https:\/\/minecraft\.azureedge\.net\/bin-win\/(bedrock-server-(.*?)\.zip)");
+                    Regex regex = new Regex(RegexString);
                     var matches = regex.Matches(html);
 
                     if (matches.Count <= 0)
@@ -235,12 +240,15 @@ namespace WindowsGSM.GameServer
                         {
                             if (Directory.Exists(Path.Combine(serverFilesPath, folder)))
                             {
-                                Directory.Delete(Path.Combine(serverFilesPath, folder),true);
+                                Directory.Delete(Path.Combine(serverFilesPath, folder), true);
                             }
                         }
+
                         File.Delete(Path.Combine(serverFilesPath, "bedrock_server.exe"));
-                        File.Delete(Path.Combine(serverFilesPath, "bedrock_server.pdb"));
-                        File.Delete(Path.Combine(serverFilesPath, "release-notes.txt"));
+                        if (File.Exists(Path.Combine(serverFilesPath, "bedrock_server.pdb")))
+                            File.Delete(Path.Combine(serverFilesPath, "bedrock_server.pdb"));
+                        if (File.Exists(Path.Combine(serverFilesPath, "release-notes.txt")))
+                            File.Delete(Path.Combine(serverFilesPath, "release-notes.txt"));
                     });
 
                     //Move folder and files
@@ -253,9 +261,12 @@ namespace WindowsGSM.GameServer
                                 Directory.Move(Path.Combine(serverFilesPath, "__temp", folder), Path.Combine(serverFilesPath, folder));
                             }
                         }
+
                         File.Move(Path.Combine(serverFilesPath, "__temp", "bedrock_server.exe"), Path.Combine(serverFilesPath, "bedrock_server.exe"));
-                        File.Move(Path.Combine(serverFilesPath, "__temp", "bedrock_server.pdb"), Path.Combine(serverFilesPath, "bedrock_server.pdb"));
-                        File.Move(Path.Combine(serverFilesPath, "__temp", "release-notes.txt"), Path.Combine(serverFilesPath, "release-notes.txt"));
+                        if (File.Exists(Path.Combine(serverFilesPath, "__temp", "bedrock_server.pdb")))
+                            File.Move(Path.Combine(serverFilesPath, "__temp", "bedrock_server.pdb"), Path.Combine(serverFilesPath, "bedrock_server.pdb"));
+                        if (File.Exists(Path.Combine(serverFilesPath, "__temp", "release-notes.txt")))
+                            File.Move(Path.Combine(serverFilesPath, "__temp", "release-notes.txt"), Path.Combine(serverFilesPath, "release-notes.txt"));
                     });
 
                     //Delete __temp folder
@@ -292,9 +303,10 @@ namespace WindowsGSM.GameServer
             {
                 return File.ReadAllText(versionPath);
             }
-            else { 
-                Error = $"Fail to get local build"; 
-                return string.Empty; 
+            else
+            {
+                Error = $"Fail to get local build";
+                return string.Empty;
             }
         }
 
@@ -305,7 +317,8 @@ namespace WindowsGSM.GameServer
                 using (WebClient webClient = new MCBEWebclient())
                 {
                     string html = await webClient.DownloadStringTaskAsync("https://www.minecraft.net/en-us/download/server/bedrock/");
-                    Regex regex = new Regex(@"https:\/\/minecraft\.azureedge\.net\/bin-win\/(bedrock-server-(.*?)\.zip)");
+
+                    Regex regex = new Regex(RegexString);
                     var matches = regex.Matches(html);
 
                     if (matches.Count > 0)
